requirements.txt
# LLM backends
google-generativeai>=0.7.0
ollama>=0.3.0

# Graph
networkx>=3.3

# SQL parsing / validation
sqlglot>=25.0.0

# API
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# UI
streamlit>=1.36.0
plotly>=5.22.0

# Data / config
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.1
pyyaml>=6.0.1

# Visualization
matplotlib>=3.9.0
graphviz>=0.20.3

# Evaluation helpers
func-timeout>=4.3.5

# Dev / test
pytest>=8.2.0
pytest-asyncio>=0.23.7
httpx>=0.27.0
ruff>=0.5.0
